hf_spaces/requirements.txt
scipy==1.4.1
tensorflow==2.2.0
keras==2.3.1
h5py==2.10.0
numpy==1.18.1
opencv-python-headless==4.2.0.32
Pillow

